fixing bugs

+++ bookings/views.py
@@ -1,11 +1,12 @@
 from django.shortcuts import get_object_or_404, render, redirect
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, user_passes_test
 from django.http import JsonResponse, Http404
 from .models import Booking, SeatReservation
 from .signals import send_booking_confirmation_email, send_booking_cancellation_email
 from .forms import BookingForm
 from django.contrib import messages
 from django.apps import apps
+from users.utils import is_staff
 
 
 Showtime = apps.get_model('showtimes', 'Showtime')
@@ -29,6 +30,7 @@ def create_booking(request, showtime_id):
       booking.save()
 
       # Process reserved seats from the form.
+      reserved_seats = form.cleaned_data["seats"]
       for seat in form.cleaned_data["seats"]:
           SeatReservation.objects.create(
               seat=seat, showtime=showtime, booking=booking, is_reserved=True
@@ -64,27 +66,32 @@ def delete_booking(request, booking_id):
     """
     Handles booking cancellation and seat release.
     """
-    booking = get_object_or_404(Booking, id=booking_id, user=request.user)
+    booking = get_object_or_404(Booking, id=booking_id)
+    if request.user == booking.user or is_staff(request.user):
+      if request.method == "POST":
+          # Collect seat details before deleting the booking
+          seats_details = ", ".join([f"{seat.seat.row_letter}{seat.seat.seat_number}" for seat in booking.seat_reservations.all()])
+          
+          # Release the reserved seats for the booking being deleted.
+          SeatReservation.objects.filter(booking=booking).update(is_reserved=False)
+          
+          # Send cancellation email before deleting the booking to have access to seat reservations
+          send_booking_cancellation_email(booking, seats_details)
 
-    if request.method == "POST":
-        # Collect seat details before deleting the booking
-        seats_details = ", ".join([f"{seat.seat.row_letter}{seat.seat.seat_number}" for seat in booking.seat_reservations.all()])
-        
-        # Release the reserved seats for the booking being deleted.
-        SeatReservation.objects.filter(booking=booking).update(is_reserved=False)
-        
-        # Send cancellation email before deleting the booking to have access to seat reservations
-        send_booking_cancellation_email(booking, seats_details)
+          # Delete the booking after sending the email
+          booking.delete()
 
-        # Delete the booking after sending the email
-        booking.delete()
-
-        messages.success(request, "Your booking has been cancelled successfully.")
-        return redirect("profile")
+          messages.success(request, "Your booking has been cancelled successfully.")
+          # Redirect staff users back to the admin dashboard and regular users to their profile.
+          if request.user.is_staff:
+              return redirect("admin_dashboard")
+          return redirect("profile")
+      else:
+          # Display confirmation for booking deletion.
+          return render(request, "bookings/confirm_delete.html", {"booking": booking})
     else:
-        # Display confirmation for booking deletion.
-        return render(request, "bookings/confirm_delete.html", {"booking": booking})
-
+      messages.error(request, "You do not have permission to cancel this booking.")
+      return redirect("profile")
 
 def seat_availability(request, showtime_id):
   """
